Rename files back
