+++ lib/yandex_mapkit.dart
@@ -48,4 +48,4 @@ part 'src/yandex_map.dart';
 part 'src/yandex_map_controller.dart';
 part 'src/yandex_search.dart';
 part 'src/yandex_search_session.dart';
-part 'src/yandex_search2.dart';
+part 'src/yandex_suggest.dart';
